+++ docs/docs-components/ColorPalette.js
@@ -1,46 +1,47 @@
 // @flow strict
-import { type Node as ReactNode } from 'react';
+import { Fragment, type Node as ReactNode } from 'react';
 import { Box, Text } from 'gestalt';
-import tokens from 'gestalt-design-tokens/dist/json/variables.json';
 import ColorTile from './ColorTile';
 
 type Props = {
   name: string,
   tokenId: string,
+  tokenData: $ReadOnlyArray<string>,
 };
 
-function ColorPalette({ name, tokenId }: Props): ReactNode {
-  const tokenNumbers = [0, 50, 100, 200, 300, 400, 500, 550, 600, 700, 800, 900];
-  const colorId = `${tokenId}-${name.toLowerCase()}`;
+function ColorPalette({ name, tokenId, tokenData }: Props): ReactNode {
+  if (tokenData) {
+    const tiles = tokenData.map((token) => {
+      const regex = /\d+(?=\D*$)/;
+      const grade = (token.match(regex) || [])[0];
+      const isTransparent = tokenId === 'transparent';
+      const textColor = Number(grade) <= 400 || isTransparent ? 'dark' : 'light';
 
-  const isTransparent = tokenId === 'transparent';
+      if (isTransparent || (grade !== '450' && !Number.isNaN(Number(grade)))) {
+        return (
+          <ColorTile
+            key={token}
+            description={grade || 'transparent'}
+            number={Number(grade) || 0}
+            textColor={textColor}
+            token={token}
+          />
+        );
+      }
 
-  return (
-    <Box marginBottom={8} marginTop={8}>
-      <Text weight="bold">
-        {name} ({tokenId})
-      </Text>
-      {isTransparent ? (
-        <Box marginTop={2}>
-          <ColorTile description="" fullTokenName="color-transparent" number={0} />
+      return null;
+    });
+
+    return (
+      <Fragment>
+        <Box marginBottom={8} marginTop={8}>
+          <Text weight="bold">
+            {name} ({tokenId})
+          </Text>
         </Box>
-      ) : (
-        <Box marginTop={2}>
-          {tokenNumbers.map((number) => {
-            const textColor = number > 400 ? 'light' : 'dark';
-            const colorVariableName = `color-${colorId}-${number}`;
-            return tokens[colorVariableName] ? (
-              <ColorTile
-                description={`${number}`}
-                fullTokenName={colorVariableName}
-                number={number}
-                textColor={textColor}
-              />
-            ) : null;
-          })}
-        </Box>
-      )}
-    </Box>
-  );
+        {tiles}
+      </Fragment>
+    );
+  }
 }
 export default ColorPalette;