+++ docs/docs-components/contexts/DocsExperimentProvider.js
@@ -10,17 +10,7 @@ import { useAppContext } from '../appContext';
  * */
 
 const enabledExperiments = {
-  ComboBox: ['web_gestalt_popover_v2_combobox', 'mweb_gestalt_popover_v2_combobox'],
-  HelpButton: ['web_gestalt_popover_v2_helpbutton', 'mweb_gestalt_popover_v2_helpbutton'],
-  OverlayPanel: [
-    'web_gestalt_popover_v2_confirmationpopover',
-    'mweb_gestalt_popover_v2_confirmationpopover',
-  ],
   Popover: ['web_gestalt_popover_v2', 'mweb_gestalt_popover_v2'],
-  PopoverEducational: [
-    'web_gestalt_popover_v2_popovereducational',
-    'mweb_gestalt_popover_v2_popovereducational',
-  ],
   Tooltip: ['web_gestalt_tooltip_v2', 'mweb_gestalt_tooltip_v2'],
 };