+++ docs/pages/web/popovereducational.js
@@ -1,7 +1,6 @@
 // @flow strict
 import { type Node as ReactNode } from 'react';
 import AccessibilitySection from '../../docs-components/AccessibilitySection';
-import { BannerSlimExperiment } from '../../docs-components/BannerSlimExperiment';
 import docGen, { type DocGen } from '../../docs-components/docgen';
 import GeneratedPropTable from '../../docs-components/GeneratedPropTable';
 import LocalizationSection from '../../docs-components/LocalizationSection';
@@ -24,17 +23,7 @@ import zIndex from '../../examples/popovereducational/zIndex';
 export default function DocsPage({ generatedDocGen }: { generatedDocGen: DocGen }): ReactNode {
   return (
     <Page title={generatedDocGen?.displayName}>
-      <PageHeader
-        bannerSlimExperiment={
-          <BannerSlimExperiment
-            componentName="PopoverEducational"
-            description="fix and improve underlying Popover component behavior. No visual updates"
-            pullRequest={3244}
-          />
-        }
-        description={generatedDocGen?.description}
-        name={generatedDocGen?.displayName}
-      >
+      <PageHeader description={generatedDocGen?.description} name={generatedDocGen?.displayName}>
         <SandpackExample code={main} hideEditor name="Main PopoverEducational example" />
       </PageHeader>