+++ packages/gestalt/src/ComboBox.js
@@ -19,12 +19,11 @@ import ComboBoxItem, { type ComboBoxItemType } from './ComboBox/Item';
 import { useDefaultLabelContext } from './contexts/DefaultLabelProvider';
 import { DOWN_ARROW, ENTER, ESCAPE, TAB, UP_ARROW } from './keyCodes';
 import Layer from './Layer';
-import Popover from './Popover';
+import InternalPopover from './Popover/InternalPopover';
 import Tag from './Tag';
 import Text from './Text';
 import InternalTextField from './TextField/InternalTextField';
 import InternalTextFieldIconButton from './TextField/InternalTextFieldIconButton';
-import useInExperiment from './useInExperiment';
 import handleContainerScrolling, {
   type DirectionOptionType,
   KEYS,
@@ -433,11 +432,6 @@ const ComboBoxWithForwardRef: AbstractComponent<Props, HTMLInputElement> = forwa
     ],
   );
 
-  const isInExperiment = useInExperiment({
-    webExperimentName: 'web_gestalt_popover_v2_combobox',
-    mwebExperimentName: 'mweb_gestalt_popover_v2_combobox',
-  });
-
   return (
     <Fragment>
       <Box
@@ -503,22 +497,21 @@ const ComboBoxWithForwardRef: AbstractComponent<Props, HTMLInputElement> = forwa
       </Box>
       {showOptionsList && innerRef.current ? (
         <Layer zIndex={zIndex}>
-          <Popover
-            key={isInExperiment ? undefined : suggestedOptions.length}
-            __experimentalPopover={isInExperiment}
+          <InternalPopover
             anchor={innerRef.current}
+            color="white"
             disablePortal
+            hideWhenReferenceHidden
             idealDirection="down"
             onDismiss={handleOnDismiss}
             onKeyDown={handleKeyDown}
-            positionRelativeToAnchor={false}
+            role="listbox"
             shouldFocus={false}
             size="flexible"
           >
             <Box
               ref={dropdownRef}
               alignItems="center"
-              aria-expanded={showOptionsList}
               direction="column"
               display="flex"
               flex="grow"
@@ -526,7 +519,6 @@ const ComboBoxWithForwardRef: AbstractComponent<Props, HTMLInputElement> = forwa
               maxHeight="30vh"
               overflow="auto"
               padding={2}
-              role="listbox"
               rounding={4}
               width={innerRef?.current?.offsetWidth}
             >
@@ -540,7 +532,7 @@ const ComboBoxWithForwardRef: AbstractComponent<Props, HTMLInputElement> = forwa
                 </Box>
               )}
             </Box>
-          </Popover>
+          </InternalPopover>
         </Layer>
       ) : null}
     </Fragment>