+++ packages/gestalt/src/OverlayPanel.jsdom.test.js
@@ -4,7 +4,6 @@ import * as AnimationControllerModule from './animation/AnimationContext';
 import { ESCAPE } from './keyCodes';
 import OverlayPanel from './OverlayPanel';
 import * as useReducedMotionHook from './useReducedMotion';
-import renderWithExperiment from './utils/testing/renderWithExperiment';
 
 jest.mock('./useReducedMotion.js');
 
@@ -23,8 +22,7 @@ describe('OverlayPanel', () => {
 
   it('should render all props with nodes', () => {
     // Remove experiment after OverlayPanel v2 is fully released
-    const { container } = renderWithExperiment(
-      'web_gestalt_popover_v2_confirmationpopover',
+    const { container } = render(
       <OverlayPanel
         accessibilityDismissButtonLabel="Dismiss"
         accessibilityLabel="OverlayPanel"