+++ playwright/accessibility/team_support__design_contributions__contribution_types_and_criteria.spec.mjs
@@ -1,0 +1,10 @@
+// @flow strict
+import { test } from '@playwright/test';
+import expectAccessiblePage from './expectAccessiblePage.mjs';
+
+test('Design handoff Accessibility check', async ({ page }) => {
+  await page.goto(
+    '/team_support/design_contributions/contribution_types_and_criteria'
+  );
+  await expectAccessiblePage({ page });
+});